fix: fixed a bug regarding updating/deleting an entity

+++ user.py
@@ -67,6 +67,13 @@ def update_user(connection, input_attribute, user_username, input_username):
     print("\nUpdating user...")
     try:
         cursor = connection.cursor()
+
+        # First, we validate if the user exists to ensure that we are not updating nothing.
+        cursor.execute("SELECT user_username FROM user WHERE user_username = %s;", (user_username,))
+        if cursor.fetchone() is None:
+            print("\nUser '{}' does not exist.\n".format(user_username))
+            return # Return if user is non-existent.
+        
         cursor.execute("UPDATE user SET {} = %s WHERE user_username = %s;".format(
             input_attribute # Attribute to be set
         ), (
@@ -78,7 +85,6 @@ def update_user(connection, input_attribute, user_username, input_username):
         # Print update details:
         print("\nUser user_username updated from '{}' to '{}' successfully!\n".format(user_username, input_username))
 
-        
     except mysql.connector.Error as err:
         print("\nError: ", err)
         print("Failed to update user.")
@@ -88,6 +94,14 @@ def delete_user(connection, user_username):
     print("\nDeleting user...")
     try:
         cursor=connection.cursor()
+
+        # Validate if user exists
+        cursor.execute("SELECT user_username FROM user WHERE user_username = %s;", (user_username,))
+        if cursor.fetchone() is None:
+            print("\nUser '{}' does not exist.\n".format(user_username))
+            return
+        
+        # Delete statement
         cursor.execute("DELETE from user WHERE user_username= %s;", (user_username,))
         rows_affected = cursor.rowcount  # Get the number of rows affected by the delete operation
         connection.commit()
